XOP-497:NetScaler suppack install fails for drivers
1. supplementalpack.py setup unconditionally sets the install (-i)
option for driver-rpms.
2. When the kernel version is not changed, but a new version of the
driver is released, the install script tries to install the new version
over the existing files. rpm reports conflicts with file due to this.
3. This commit introduces a new option (-m) for use with kernel modules.
This is used for packages that have kernel version in their name (like
driver-rpm). (-i) will continue to be used for kernel packages that do
not include version in name (like kernel, kernel-dump).
4. During install, if -m in options, choose to upgrade when module with
same name is already installed, else install.